make check_integral_table 75% faster by using eq_with_var instead of replace
